Add TypeScript entry point
